Added skill damageModifier when a fighter attack, Test 4 Passed

+++ src/main/java/me/guillaume/tournament/Veteran.java
@@ -5,13 +5,12 @@ public class Veteran extends Skill {
     private static final int BERSERK_HEALTH_PERCENTAGE_THRESHOLD_TRIGGER = 30;
     private static final int BERSERK_DAMAGE_MULTIPLIER = 2;
 
-
     public Veteran(){
     }
 
     @Override
     public int damageModifier(int attackPower, Fighter fighter) {
-        int fighterHealthPercentage = (fighter.hitPoints() / fighter.maxHitPoints()) * 100;
+        float fighterHealthPercentage = ((float) fighter.hitPoints() / (float) fighter.maxHitPoints()) * 100;
 
         if(fighterHealthPercentage < BERSERK_HEALTH_PERCENTAGE_THRESHOLD_TRIGGER){
             return attackPower * BERSERK_DAMAGE_MULTIPLIER;
